restructure from rocket-chip to radiance
