+++ build.sbt
@@ -1,0 +1,7 @@
+version := "0.1"
+
+name := "radiance"
+
+organization := "edu.berkeley.cs"
+
+scalaVersion := "2.13.10"
