+++ src/main/scala/radiance/memory/VortexBank.scala
@@ -1,4 +1,4 @@
-package freechips.rocketchip.tilelink
+package radiance.memory
 
 import chisel3._
 import chisel3.util._
