+++ src/main/scala/radiance/tile/VortexCore.scala
@@ -1,7 +1,7 @@
 // See LICENSE.Berkeley for license details.
 // See LICENSE.SiFive for license details.
 
-package rocket
+package radiance.tile
 
 import chisel3._
 import chisel3.util._
