--- src/main/scala/subsystem/RocketSubsystem.scala
@@ -1,43 +1,0 @@
-// See LICENSE.SiFive for license details.
-
-package freechips.rocketchip.subsystem
-
-import org.chipsalliance.cde.config.Parameters
-import freechips.rocketchip.diplomacy._
-import freechips.rocketchip.prci.{NoResetCrossing, ResetCrossingType}
-import freechips.rocketchip.tile._
-import freechips.rocketchip.devices.debug.HasPeripheryDebug
-
-case class RocketCrossingParams(
-  crossingType: ClockCrossingType = SynchronousCrossing(),
-  master: TilePortParamsLike = TileMasterPortParams(),
-  slave: TileSlavePortParams = TileSlavePortParams(),
-  mmioBaseAddressPrefixWhere: TLBusWrapperLocation = CBUS,
-  resetCrossingType: ResetCrossingType = NoResetCrossing(),
-  forceSeparateClockReset: Boolean = false
-) extends TileCrossingParamsLike
-
-case class RocketTileAttachParams(
-  tileParams: RocketTileParams,
-  crossingParams: RocketCrossingParams
-) extends CanAttachTile { type TileType = RocketTile }
-
-case class VortexTileAttachParams(
-  tileParams: VortexTileParams,
-  crossingParams: RocketCrossingParams
-) extends CanAttachTile { type TileType = VortexTile }
-
-trait HasRocketTiles extends HasTiles { this: BaseSubsystem =>
-  val rocketTiles = tiles.collect { case r: RocketTile => r }
-
-  def coreMonitorBundles = (rocketTiles map { t =>
-    t.module.core.rocketImpl.coreMonitorBundle
-  }).toList
-}
-
-class RocketSubsystem(implicit p: Parameters) extends BaseSubsystem with HasRocketTiles with HasPeripheryDebug {
-  override lazy val module = new RocketSubsystemModuleImp(this)
-}
-
-class RocketSubsystemModuleImp[+L <: RocketSubsystem](_outer: L) extends BaseSubsystemModuleImp(_outer)
-    with HasTilesModuleImp
